parse_node: move contents to forest.
