+++ src/forest.rs
@@ -1,12 +1,44 @@
 use crate::high::{type_lambda, ExistsL, PairL};
 use crate::input::{Input, Range};
-use crate::parse_node::ParseNodeShape;
 use indexing::{self, Container};
 use std::collections::{BTreeSet, HashMap, VecDeque};
 use std::fmt;
 use std::hash::Hash;
 use std::io::{self, Write};
 use std::str;
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum ParseNodeShape<P> {
+    Opaque,
+    Alias(P),
+    Choice,
+    Opt(P),
+    Split(P, P),
+}
+
+impl<P: fmt::Display> fmt::Display for ParseNodeShape<P> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ParseNodeShape::Opaque => write!(f, "Opaque"),
+            ParseNodeShape::Alias(inner) => write!(f, "Alias({})", inner),
+            ParseNodeShape::Choice => write!(f, "Choice"),
+            ParseNodeShape::Opt(inner) => write!(f, "Opt({})", inner),
+            ParseNodeShape::Split(left, right) => write!(f, "Split({}, {})", left, right),
+        }
+    }
+}
+
+impl<P> ParseNodeShape<P> {
+    pub fn map<Q>(self, mut f: impl FnMut(P) -> Q) -> ParseNodeShape<Q> {
+        match self {
+            ParseNodeShape::Opaque => ParseNodeShape::Opaque,
+            ParseNodeShape::Alias(inner) => ParseNodeShape::Alias(f(inner)),
+            ParseNodeShape::Choice => ParseNodeShape::Choice,
+            ParseNodeShape::Opt(inner) => ParseNodeShape::Opt(f(inner)),
+            ParseNodeShape::Split(left, right) => ParseNodeShape::Split(f(left), f(right)),
+        }
+    }
+}
 
 /// Objects capable of providing information about various parts of the grammar
 /// (mostly parse nodes and their substructure).
